sync(official): EML, PIE, EMUI9.0

+++ drivers/huawei_platform/nfc/cxd224x/cxd224x-i2c.c
@@ -457,7 +457,7 @@ static long cxd224x_dev_unlocked_ioctl(struct file *filp,
 }
 
 
-static int cxd224x_pon_low_beforepwd(struct notifier_block *this, unsigned int code, void *unused)
+static int cxd224x_pon_low_beforepwd(struct notifier_block *this, unsigned long code, void *unused)
 {
 	int retval = 0;
